Remove parens
