+++ webpack/account/labs/__tests__/fetch_lab_features_test.ts
@@ -1,0 +1,15 @@
+import { fetchLabFeatures } from "../labs_features_list_data";
+
+describe("fetchLabFeatures", () => {
+  it("basically just initializes stuff", () => {
+    const val = fetchLabFeatures();
+    expect(val.length).toBe(6);
+    expect(val[0].value).toBeFalsy();
+    const { callback } = val[0];
+    if (callback) {
+      expect(callback).not.toThrowError();
+    } else {
+      expect(callback).toBeDefined();
+    }
+  });
+});
